requirements.txt
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn>=0.27.0
pandas>=2.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
plotly>=5.18.0
watchdog>=3.0.0
openai>=1.10.0
requests>=2.31.0
sqlmodel>=0.0.14
